Rename servicecheck package
