+++ eagle-hadoop-servicecheck/src/main/resources/META-INF/services/org.apache.eagle.app.spi.ApplicationProvider
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-org.apache.eagle.healthcheck.HadoopServiceCheckAppProvider
+servicecheck.HadoopServiceCheckAppProvider
